feat: add sessionCard component
