+++ packages/core/src/components/organisms/index.js
@@ -27,6 +27,7 @@ export {default as MailMessageCommentCard} from './MailMessageCommentCard/MailMe
 export {default as MailMessageNotificationCard} from './MailMessageNotificationCard/MailMessageNotificationCard';
 export {default as PasswordInput} from './PasswordInput/PasswordInput';
 export {default as ScannerAutocompleteSearch} from './ScannerAutocompleteSearch/ScannerAutocompleteSearch';
+export {default as SessionCard} from './SessionCard/SessionCard';
 export {default as SessionNameInput} from './SessionNameInput/SessionNameInput';
 export {default as UrlInput} from './UrlInput/UrlInput';
 export {default as UsernameInput} from './UsernameInput/UsernameInput';
