+++ packages/core/src/components/templates/SessionListCard/SessionListCard.js
@@ -16,29 +16,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, {useCallback, useMemo, useRef, useState} from 'react';
-import {
-  ScrollView,
-  StyleSheet,
-  TouchableOpacity,
-  View,
-  Dimensions,
-  Animated,
-} from 'react-native';
-import {
-  useThemeColor,
-  Card,
-  Text,
-  InfoBubble,
-  Icon,
-  useSwipe,
-  SquareIcon,
-} from '@axelor/aos-mobile-ui';
+import React, {useMemo} from 'react';
+import {ScrollView, StyleSheet, View, Dimensions} from 'react-native';
+import {useThemeColor, InfoBubble, Icon} from '@axelor/aos-mobile-ui';
 import {useTranslator} from '../../../i18n';
-import {sessionStorage} from '../../../sessions';
-import {LogoImage} from '../../organisms';
+import {SessionCard} from '../../organisms';
 import {SessionNumberIndicator} from '../../molecules';
-import {useEffect} from 'react';
 
 const SessionListCard = ({
   sessionList,
@@ -53,111 +36,12 @@ const SessionListCard = ({
   const Colors = useThemeColor();
   const I18n = useTranslator();
 
-  const {onTouchStart, onTouchEnd} = useSwipe(onSwipeLeft, onSwipeRight, 6);
-
-  const [showButton, setShowButton] = useState(false);
-
   const sessions = useMemo(() => sessionList, [sessionList]);
 
   const styles = useMemo(() => getStyles(Colors), [Colors]);
 
-  const translateXAnim = useRef(new Animated.Value(0)).current;
-  const translateYAnim = useRef(new Animated.Value(0)).current;
-  const scaleAnim = useRef(new Animated.Value(1)).current;
-
   const selectedIndex = sessions?.findIndex(
     _session => _session?.sessionId === session?.sessionId,
-  );
-
-  useEffect(() => {
-    if (showButton) {
-      Animated.timing(translateXAnim, {
-        toValue: -60,
-        duration: 200,
-        useNativeDriver: true,
-      }).start();
-    } else {
-      Animated.timing(translateXAnim, {
-        toValue: 0,
-        duration: 500,
-        useNativeDriver: true,
-      }).start();
-    }
-  }, [
-    setAuthorizePopupToOpen,
-    setPopupSessionIsOpen,
-    showButton,
-    translateXAnim,
-  ]);
-
-  function onSwipeLeft() {
-    setAuthorizePopupToOpen(false);
-    setShowButton(true);
-    setPopupSessionIsOpen(false);
-  }
-
-  function onSwipeRight() {
-    setShowButton(false);
-    setAuthorizePopupToOpen(false);
-    setPopupSessionIsOpen(false);
-  }
-
-  const removeSession = useCallback(sessionId => {
-    sessionStorage.removeSession({sessionId});
-  }, []);
-
-  const changeActiveSession = useCallback(
-    sessionId => {
-      sessionStorage.changeActiveSession({sessionId});
-      onChange(sessionStorage.getActiveSession());
-    },
-    [onChange],
-  );
-
-  const changeActiveSessionPopup = useCallback(
-    sessionId => {
-      setAuthorizePopupToOpen(true);
-      sessionStorage.changeActiveSession({sessionId});
-      onChange(sessionStorage.getActiveSession());
-      setPopupSessionIsOpen(true);
-    },
-    [setAuthorizePopupToOpen, onChange, setPopupSessionIsOpen],
-  );
-
-  const animateRemoval = useCallback(
-    index => {
-      setAuthorizePopupToOpen(false);
-      Animated.parallel([
-        Animated.timing(scaleAnim, {
-          toValue: 0,
-          duration: 500,
-          useNativeDriver: true,
-        }),
-        Animated.timing(translateXAnim, {
-          toValue: Dimensions.get('window').width * 0.5,
-          duration: 500,
-          useNativeDriver: true,
-        }),
-        Animated.timing(translateYAnim, {
-          toValue: Dimensions.get('window').height * 0.3,
-          duration: 500,
-          useNativeDriver: true,
-        }),
-      ]).start(() => {
-        removeSession(sessionList[index].sessionId);
-        scaleAnim.setValue(1);
-        translateXAnim.setValue(0);
-        translateYAnim.setValue(0);
-      });
-    },
-    [
-      setAuthorizePopupToOpen,
-      removeSession,
-      sessionList,
-      scaleAnim,
-      translateXAnim,
-      translateYAnim,
-    ],
   );
 
   if (!Array.isArray(sessions) || sessions?.length === 0) {
@@ -188,70 +72,19 @@ const SessionListCard = ({
         style={styles.scrollView}>
         {sessions.map((_session, index) => {
           return (
-            <View key={index} style={styles.listContainer}>
-              <TouchableOpacity
-                onLongPress={() => {
-                  changeActiveSession(_session.sessionId);
-                  onSwipeLeft();
-                }}
-                onPress={() => {
-                  onSwipeRight();
-                  changeActiveSessionPopup(_session.sessionId);
-                }}
-                activeOpacity={0.9}>
-                <Animated.View
-                  onTouchStart={onTouchStart}
-                  onTouchEnd={onTouchEnd}
-                  style={[
-                    selectedIndex === index
-                      ? {
-                          transform: [
-                            {translateX: translateXAnim},
-                            {translateY: translateYAnim},
-                            {scale: scaleAnim},
-                          ],
-                        }
-                      : null,
-                  ]}>
-                  <Card
-                    style={[
-                      styles.cardContainer,
-                      _session?.sessionId === session?.sessionId
-                        ? styles.selectBorderCard
-                        : styles.borderCard,
-                      _session?.isDefault ? styles.borderIsDefaultCard : null,
-                    ]}>
-                    <View style={styles.imageContainer}>
-                      <LogoImage logoFile={logoFile} url={_session?.url} />
-                    </View>
-                    <View style={styles.textContainer}>
-                      <Text style={styles.textTitle}>{_session.id}</Text>
-                      <Text
-                        numberOfLines={1}
-                        style={styles.details}>{`Url: ${_session.url}`}</Text>
-                    </View>
-                  </Card>
-                </Animated.View>
-              </TouchableOpacity>
-              {showButton && _session?.sessionId === session?.sessionId && (
-                <Animated.View
-                  style={[
-                    styles.squareIconContainer,
-                    {transform: [{translateX: translateXAnim}]},
-                  ]}>
-                  <SquareIcon
-                    iconName="pencil-alt"
-                    borderColor={Colors.infoColor.background}
-                    onPress={() => setPopupEditIsOpen(true)}
-                  />
-                  <SquareIcon
-                    iconName="trash-alt"
-                    borderColor={Colors.errorColor.background}
-                    onPress={() => animateRemoval(index)}
-                  />
-                </Animated.View>
-              )}
-            </View>
+            <SessionCard
+              key={index}
+              _session={_session}
+              index={index}
+              logoFile={logoFile}
+              onChange={onChange}
+              selectedIndex={selectedIndex}
+              session={session}
+              sessionList={sessionList}
+              setAuthorizePopupToOpen={setAuthorizePopupToOpen}
+              setPopupEditIsOpen={setPopupEditIsOpen}
+              setPopupSessionIsOpen={setPopupSessionIsOpen}
+            />
           );
         })}
       </ScrollView>
@@ -264,47 +97,12 @@ const getStyles = Colors =>
     container: {
       maxHeight: Dimensions.get('window').height * 0.45,
     },
-    listContainer: {
-      flexDirection: 'row',
-      alignItems: 'center',
-    },
     scrollView: {
       width: '100%',
     },
     contentContainer: {
       alignItems: 'center',
       paddingHorizontal: '5%',
-    },
-    cardContainer: {
-      flexDirection: 'row',
-      marginVertical: 4,
-      width: '100%',
-      alignItems: 'center',
-      height: Dimensions.get('window').height * 0.15,
-    },
-    borderCard: {
-      borderWidth: 1,
-      borderColor: Colors.secondaryColor_dark.background,
-    },
-    selectBorderCard: {
-      borderWidth: 1,
-    },
-    borderIsDefaultCard: {
-      borderWidth: 1,
-      borderColor: Colors.primaryColor.background,
-    },
-    textTitle: {
-      fontSize: 18,
-      fontWeight: 'bold',
-    },
-    textContainer: {
-      flexDirection: 'column',
-      width: '70%',
-      marginLeft: 20,
-    },
-    imageContainer: {
-      width: '20%',
-      height: 100,
     },
     iconContainer: {
       alignSelf: 'flex-end',
@@ -328,11 +126,6 @@ const getStyles = Colors =>
       borderRadius: Dimensions.get('window').width * 0.07,
       width: Dimensions.get('window').width * 0.07,
       height: Dimensions.get('window').width * 0.07,
-    },
-    squareIconContainer: {
-      flexDirection: 'column',
-      position: 'absolute',
-      right: -Dimensions.get('window').height * 0.07 - 10,
     },
     textIndicationStyle: {
       width: Dimensions.get('window').width * 0.7,
